implement car waypoints tracking
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@1481 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/game.h
@@ -73,7 +73,7 @@ public:
 	
 	const std::string getRandomWaypoint(const std::string &classname, const std::string &last_wp = std::string()) const;
 	const std::string getNearestWaypoint(const BaseObject *obj, const std::string &classname) const;
-	void getWaypoint(v3<int> &wp, const std::string &classname, const std::string &name);
+	void getWaypoint(v3<float> &wp, const std::string &classname, const std::string &name);
 	
 private:
 	void onKey(const Uint8 type, const SDL_keysym sym);
